+++ src/solys2/__init__.py
@@ -2,3 +2,4 @@ from . import solys2
 from . import connection
 from . import automation
 from . import response
+from . import common