+++ src/solys2/automation/__init__.py
@@ -1,3 +1,4 @@
 from . import autohelper
 from . import autotrack
 from . import calibration
+from . import positioncalc